worked on player stats. worked on artillery class. worked on player progression.

+++ demo/src/main/java/com/cycapservers/game/Entity.java
@@ -42,8 +42,8 @@ public class Entity {
 		output += spriteIndex + ",";
 		output += (int) x + ",";
 		output += (int) y + ",";
-		output += (int) drawWidth + ",";
-		output += (int) drawHeight + ",";
+		output += drawWidth + ",";
+		output += drawHeight + ",";
 		output += rotation + ",";
 		output += alpha;
 		return output;
